10k rows flag

+++ backend/alembic/versions/c83affcaea8a_populate_10000_rows.py
@@ -8,6 +8,7 @@ Create Date: 2022-03-04 17:44:48.478986
 from alembic import op
 import sqlalchemy as sa
 from generated import generated
+import os
 
 # revision identifiers, used by Alembic.
 revision = 'c83affcaea8a'
@@ -17,6 +18,7 @@ depends_on = None
 
 
 def upgrade():
+  if not os.getenv('10K_ROWS'):
     for player in generated:
 
       command = 'insert into rushing_stats ("player","team","pos","att","att_g","yds","avg","yds_g","td","lng","first","first_percent","twenty_plus","forty_plus","fum") \
@@ -30,6 +32,7 @@ def upgrade():
 
 
 def downgrade():
+  if not os.getenv('10K_ROWS'):
     for player in generated:
       command = 'delete from rushing_stats where player=\'{}\' and  team=\'{}\' and pos=\'{}\''.format(str(player['player']).replace("'","''"), player['team'], player['pos'])
       op.execute(command)
